refactor: clean up project
